pyproject.toml
[tool.poetry]
name = "zoomdashboard"
version = "0.1.0"
description = ""
authors = ["tsudo <tsudonymn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
assertpy = "^1.1"
pandas = "^2.2.3"
dash = "^2.18.2"
dash-bootstrap-components = "^1.5.0"
paho-mqtt = "1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
